tests/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
blinker==1.9.0
certifi==2025.6.15
click==8.1.8
cloudflare==4.3.1
coverage==7.9.2
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pip==25.1
platformdirs==4.3.8
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
setuptools==78.1.1
sniffio==1.3.1
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.1
typing-inspection==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
zipp==3.23.0
